requirements.txt
click==7.1.2
python-git-info==0.6.1
motor==2.3.0
pynacl==1.4.0
tornado==6.0.4
